requirements.txt
pandas
yfinance
selenium
python-dotenv